+++ pTyping/Graphics/Drawables/Events/TypingCutoffEventDrawable.cs
@@ -27,7 +27,7 @@ namespace pTyping.Graphics.Drawables.Events {
             Vector2 recepticlePos = tweenArgs.IsEditor ? EditorScreen.RECEPTICLE_POS : Player.Player.RECEPTICLE_POS;
 
             float travelDistance = noteStartPos.X - recepticlePos.X;
-            float travelRatio    = tweenArgs.ApproachTime / travelDistance;
+            float travelRatio    = (float)(tweenArgs.ApproachTime / travelDistance);
 
             float afterTravelTime = (recepticlePos.X - noteEndPos.X) * travelRatio;
